refactor: 폼 제출 로직 분리
